meal_prep_app/serializers.py
from django.contrib.auth.models import User, Group
from .models import Recipe

from rest_framework import serializers


class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ['url', 'username', 'email', 'groups']


class RecipeSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Recipe
        fields = ['title', 'pub_date', 'description', 'recipe']
