requirements.txt
asgiref==3.3.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
distlib==0.3.2
Django==3.1.7
django-debug-toolbar==3.2.1
django-multiselectfield==0.1.12
django-pluralize-pl==0.2.1
djangorestframework==3.12.2
filelock==3.0.12
importlib-metadata==4.6.1
importlib-resources==5.2.0
pipenv==2021.5.29
pkg-resources==0.0.0
platformdirs==2.0.2
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
virtualenv==20.5.0
virtualenv-clone==0.5.4
zipp==3.5.0
